Add PLP and search widget
