+++ blocks/product-list-page/product-list-page.js
@@ -1,279 +1,55 @@
-import {
-  h, Component, Fragment, render, createRef,
-} from '../../scripts/preact.js';
-import htm from '../../scripts/htm.js';
-import ProductList from './ProductList.js';
-import FacetList from './FacetList.js';
-import { readBlockConfig } from '../../scripts/aem.js';
-import {
-  loadCategory,
-  parseQueryParams,
-  ALLOWED_FILTER_PARAMETERS,
-} from '../../scripts/commerce.js';
-
-const html = htm.bind(h);
-
-function Pagination(props) {
-  if (props.loading) {
-    return html`<div class="pagination shimmer"></div>`;
-  }
-
-  return html`<div class="pagination">
-    <div>
-      <label for="select-pagesize">Show:</label>
-      <select id="select-pagesize" name="pageSize" value=${props.currentPageSize} onChange=${(e) => props.onPageSizeChange?.(parseInt(e.target.value, 10))}>
-        ${props.pageSizeOptions.map((size) => html`<option value=${size}>${size} Items</option>`)}
-      </select>
-    </div>
-    <div>
-      <label for="select-page">Page:</label>
-      <select id="select-page" name="page" value=${props.currentPage} onChange=${(e) => props.onPageChange?.(parseInt(e.target.value, 10))}>
-        ${Array(props.pages).fill(0).map((_, i) => html`<option value=${i + 1}>${i + 1}</option>`)}
-      </select>
-      <span>of ${props.pages}</span>
-    </div>
-    <div>
-      ${props.currentPage > 1 ? html`<button class="previous" onClick=${() => props.onPageChange?.(props.currentPage - 1)}>Previous</button>` : ''}
-      ${props.currentPage < props.pages ? html`<button class="next" onClick=${() => props.onPageChange?.(props.currentPage + 1)}>Next</button>` : ''}
-    </div>
-  </div>`;
-}
-
-function Sort(props) {
-  const {
-    type, disabled, sortMenuRef, onSort,
-  } = props;
-  const options = [
-    { label: 'Price: High to Low', value: 'price-desc' },
-    { label: 'Price: Low to High', value: 'price-asc' },
-    { label: 'Product Name', value: 'name-asc' },
-    { label: 'Relevance', value: type === 'category' ? 'position-asc' : 'relevance-desc' },
-  ];
-
-  const currentSort = options.find((option) => option.value === `${props.currentSort}-${props.sortDirection}`) || options[3];
-
-  return html`<div class="sort" disabled=${disabled}>
-    <button disabled=${disabled}>Sort By: ${currentSort.label}</button>
-    <div class="overlay" ref=${sortMenuRef}>
-      <button class="close" onClick=${() => sortMenuRef.current.classList.toggle('active')}>Close</button>
-      <ul>
-        ${options.map((option) => html`<li>
-          <a href="#" class="${currentSort.value === option.value ? 'active' : ''}" onClick=${(e) => {
-  sortMenuRef.current.classList.toggle('active');
-  const [sort, direction = 'asc'] = option.value.split('-');
-  onSort?.(sort, direction);
-  e.preventDefault();
-}}>${option.label}</a>
-        </li>`)}
-      </ul>
-    </div>
-  </div>`;
-}
-
-class ProductListPage extends Component {
-  constructor(props) {
-    const {
-      type = 'category',
-      category,
-    } = props;
-    super();
-
-    this.facetMenuRef = createRef();
-    this.sortMenuRef = createRef();
-
-    const queryParams = parseQueryParams();
-
-    let headline = 'Search Results';
-    let sort = 'relevance';
-    let sortDirection = 'desc';
-    const defaultPageSize = 12;
-    if (type === 'category') {
-      // Get from H1
-      headline = document.querySelector('.default-content-wrapper > h1')?.innerText;
-      sort = 'position';
-      sortDirection = 'asc';
-    }
-
-    /* if (type === 'search') {
-      sampleRUM('search', { source: '.search-input', target: queryParams.searchTerm });
-    } */
-
-    this.state = {
-      loading: true,
-      pages: 1,
-      currentPage: 1,
-      basePageSize: defaultPageSize,
-      currentPageSize: defaultPageSize,
-      type,
-      category: {
-        name: headline,
-        id: category || null,
-      },
-      sort,
-      sortDirection,
-      products: {
-        items: [],
-        total: 0,
-      },
-      filters: {},
-      facets: [],
-      ...queryParams,
-    };
-
-    this.filterChange = false;
-    this.paginationClick = false;
-  }
-
-  setStatePromise(state) {
-    return new Promise((resolve) => {
-      this.setState(state, resolve);
-    });
-  }
-
-  static updateQueryParams = (params) => {
-    const newParams = new URLSearchParams();
-    Object.keys(params).forEach((key) => {
-      if (!ALLOWED_FILTER_PARAMETERS.includes(key)) {
-        return;
-      }
-
-      if (Array.isArray(params[key]) && params[key].length > 0) {
-        newParams.set(key, params[key].join(','));
-      } else if (!Array.isArray(params[key]) && params[key]) {
-        newParams.set(key, params[key]);
-      }
-    });
-
-    // Keep existing params that do not interfere with the search
-    const curentParams = new URLSearchParams(window.location.search);
-    curentParams.forEach((value, key) => {
-      if (!ALLOWED_FILTER_PARAMETERS.includes(key)) {
-        newParams.set(key, value);
-      }
-    });
-
-    window.history.pushState({}, '', `${window.location.pathname}?${newParams.toString()}`);
-  };
-
-  loadState = async (state) => {
-    await this.setStatePromise({ ...state, loading: false });
-    if (this.state && this.state.products) {
-      this.filterChange = false;
-      this.paginationClick = false;
-    }
-    this.props.resolve();
-    /* getMagentoStorefrontEvents((mse) => {
-      mse.context.setCategory(loadPageDetails());
-      mse.publish.pageView();
-    }); */
-  };
-
-  loadProducts = async () => {
-    this.setState({ loading: true });
-
-    const state = await loadCategory(this.state);
-    await this.loadState(state);
-  };
-
-  async componentDidMount() {
-    if (window.loadCategoryPromise) {
-      const state = await window.loadCategoryPromise;
-      await this.loadState(state);
-    } else {
-      await this.loadProducts();
-    }
-  }
-
-  componentDidUpdate(_, prevState) {
-    // Update URL
-    ProductListPage.updateQueryParams({
-      page: this.state.currentPage,
-      basePageSize: this.state.basePageSize,
-      pageSize: this.state.currentPageSize,
-      sort: this.state.sort,
-      sortDirection: this.state.sortDirection,
-      q: this.state.searchTerm,
-      ...this.state.filters,
-    });
-
-    // Load new products if filters, pagination or sort have changed
-    const diff = Object.keys(Object.keys(prevState).reduce((acc, key) => {
-      if (prevState[key] !== this.state[key]) {
-        acc[key] = this.state[key];
-      }
-      return acc;
-    }, {}));
-
-    const keysToCheck = ['filters', 'sort', 'sortDirection', 'searchTerm', 'currentPageSize', 'currentPage'];
-    if (keysToCheck.some((key) => diff.includes(key))) {
-      this.loadProducts();
-    }
-
-    /* if (this.props.type === 'search'
-    && this.state.loading === false && this.state.products.total === 0) {
-      sampleRUM('nullsearch', { source: '.search-input', target: this.state.searchTerm });
-    } */
-  }
-
-  onPageChange(page) {
-    this.setState({ currentPage: page });
-    this.paginationClick = true;
-    window.scrollTo({ top: 0, behavior: 'smooth' });
-  }
-
-  handleFilterChange(filters) {
-    this.setState({ filters, currentPage: 1 });
-    this.filterChange = true;
-  }
-
-  render(props, state) {
-    const { type = 'category' } = props;
-
-    return html`<${Fragment}>
-    <${FacetList} 
-      facets=${state.facets}
-      filters=${state.filters}
-      facetMenuRef=${this.facetMenuRef}
-      onFilterChange=${this.handleFilterChange.bind(this)}
-      loading=${state.loading} />
-    <div class="products">
-      <div class="title">
-        <h1>${state.category.name}</h1>
-        ${!state.loading && html`<span>(${state.products.total} ${state.products.total === 1 ? 'Product' : 'Products'})</span>`}
-        <${Sort}
-          disabled=${state.loading}
-          currentSort=${state.sort}
-          sortDirection=${state.sortDirection}
-          type=${type}
-          onSort=${(sort, direction) => this.setState({ sort, sortDirection: direction })}
-          sortMenuRef=${this.sortMenuRef} />
-      </div>
-      <div class="mobile-menu">
-        <button disabled=${state.loading} id="toggle-filters" onClick=${() => this.facetMenuRef.current.classList.toggle('active')}>Filters</button>
-        <button disabled=${state.loading} id="toggle-sortby" onClick=${() => this.sortMenuRef.current.classList.toggle('active')}>Sort By</button>
-      </div>
-      <${ProductList} products=${state.products} disruptors=${state.disruptors} loading=${state.loading} currentPageSize=${state.currentPageSize} />
-      <${Pagination}
-        pages=${state.pages}
-        currentPage=${state.currentPage}
-        pageSizeOptions=${[state.basePageSize, 24, 36]}
-        currentPageSize=${state.currentPageSize}
-        loading=${state.loading}
-        onPageChange=${this.onPageChange.bind(this)}
-        onPageSizeChange=${(pageSize) => this.setState({ currentPageSize: pageSize, currentPage: 1 })} />
-    </div>
-  </>`;
-  }
-}
+import { readBlockConfig, loadScript } from '../../scripts/aem.js';
+import { getConfigValue } from '../../scripts/configs.js';
 
 export default async function decorate(block) {
-  const config = readBlockConfig(block);
-
+  const { urlpath, category, type } = readBlockConfig(block);
   block.textContent = '';
-  block.dataset.category = config.category;
 
-  return new Promise((resolve) => {
-    const app = html`<${ProductListPage} ...${config} block=${block} resolve=${resolve} />`;
-    render(app, block);
+  const widgetProd = 'https://plp-widgets-ui.magento-ds.com/v1/search.js';
+  await loadScript(widgetProd);
+
+  const storeDetails = {
+    environmentId: await getConfigValue('commerce-environment-id'),
+    environmentType: 'testing',
+    apiKey: await getConfigValue('commerce-x-api-key'),
+    websiteCode: await getConfigValue('commerce-website-code'),
+    storeCode: await getConfigValue('commerce-store-code'),
+    storeViewCode: await getConfigValue('commerce-store-view-code'),
+    config: {
+      pageSize: 8,
+      perPageConfig: {
+        pageSizeOptions: '12,24,36',
+        defaultPageSizeOption: '24',
+      },
+      minQueryLength: '2',
+      currencySymbol: '$',
+      currencyRate: '1',
+      displayOutOfStock: true,
+      allowAllProducts: false,
+      displayMode: '', // "" for plp || "PAGE" for category/catalog
+    },
+    context: {
+      customerGroup: await getConfigValue('commerce-customer-group'),
+    },
+    route: ({ sku }) => `/products/missing-url-key/${sku}`, // TODO: We need urlKey as parameter as well!
+  };
+
+  if (type !== 'search') {
+    storeDetails.config.categoryName = document.querySelector('.default-content-wrapper > h1')?.innerText;
+    storeDetails.config.currentCategoryUrlPath = urlpath;
+
+    // Enable enrichment
+    block.dataset.category = category;
+  }
+
+  await new Promise((resolve) => {
+    const interval = setInterval(() => {
+      if (window.LiveSearchPLP) {
+        clearInterval(interval);
+        resolve();
+      }
+    }, 200);
   });
+
+  window.LiveSearchPLP({ storeDetails, root: block });
 }
